# src/main/resources/openapi/customers-api.yaml
openapi: 3.0.3
info:
  title: Bookstore Customer API
  version: 1.0.0
paths:
  /customers:
    get:
      summary: Get all customers
      operationId: getAllCustomers
      responses:
        '200':
          description: A list of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerResponse'
    post:
      summary: Create a new customer
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        '201':
          description: Customer created
          headers:
            Location:
              description: URI of the created customer
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'

  /customers/{id}:
    get:
      summary: Get customer by ID
      operationId: getCustomer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Customer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '404':
          description: Customer not found
    put:
      summary: Update customer
      operationId: updateCustomer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        '200':
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
    delete:
      summary: Delete customer
      operationId: deleteCustomer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Customer deleted

  /customers/{id}/loyalty:
    get:
      summary: Get loyalty points of a customer
      operationId: getLoyaltyPoints
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Loyalty points retrieved
          content:
            application/json:
              schema:
                type: integer
                format: int32
        '404':
          description: Customer not found

components:
  schemas:
    CustomerRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1

    CustomerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
        loyaltyPoints:
          type: integer
          format: int32
          minimum: 1
